fix(nccl): populate window in ncclDevApiMemCreate for Vendor path
