+++ flagcx/adaptor/device_api/default_dev_api_backend.cc
@@ -36,140 +36,6 @@ static void flagcxDevNetLaunchConstruct(void *, void *, int, void *) {}
 #endif
 
 // ==========================================================================
-// IPC peer pointer exchange
-// ==========================================================================
-
-int buildIpcPeerPointers(flagcxComm_t comm, void *buff, size_t size) {
-  int slot = -1;
-  for (int k = 0; k < FLAGCX_MAX_IPC_ENTRIES; k++) {
-    if (comm->ipcTable[k].hostPeerPtrs == nullptr &&
-        comm->ipcTable[k].devPeerPtrs == nullptr) {
-      slot = k;
-      break;
-    }
-  }
-  if (slot < 0) {
-    WARN("buildIpcPeerPointers: IPC table full (max %d entries)",
-         FLAGCX_MAX_IPC_ENTRIES);
-    return -1;
-  }
-
-  int myRank = comm->rank;
-  int nRanks = comm->nranks;
-  int localRanks = comm->localRanks;
-  int *localRankToRank = comm->localRankToRank;
-
-  flagcxResult_t res = flagcxSuccess;
-  struct flagcxP2pIpcDesc *allDescs = nullptr;
-  void **hostPeerPtrs = nullptr;
-  void **devPeerPtrs = nullptr;
-
-  // Step 1: Get IPC handle for existing user buffer.
-  struct flagcxP2pIpcDesc myIpcDesc;
-  memset(&myIpcDesc, 0, sizeof(myIpcDesc));
-
-  // Check globalRegPool for pre-registered handle
-  {
-    flagcxRegItem *item = globalRegPool.getItem(nullptr, buff);
-    if (item && item->localIpcHandleData.reserved[0] != 0) {
-      memcpy(&myIpcDesc.handleData, &item->localIpcHandleData,
-             sizeof(myIpcDesc.handleData));
-      myIpcDesc.size = size;
-    } else {
-      // Create IPC handle on the fly
-      flagcxP2pAllocateShareableBuffer(size, 0, &myIpcDesc, (void **)&buff);
-      myIpcDesc.size = size;
-    }
-  }
-
-  // Step 2: All-gather IPC descriptors across local ranks
-  allDescs = (struct flagcxP2pIpcDesc *)malloc(nRanks *
-                                               sizeof(struct flagcxP2pIpcDesc));
-  if (!allDescs)
-    return -1;
-  memset(allDescs, 0, nRanks * sizeof(struct flagcxP2pIpcDesc));
-  allDescs[myRank] = myIpcDesc;
-
-  FLAGCXCHECKGOTO(bootstrapCollAllGather(comm->bootstrap, allDescs,
-                                         sizeof(struct flagcxP2pIpcDesc)),
-                  res, fail);
-
-  // Step 3: Open peer IPC handles
-  hostPeerPtrs = (void **)malloc(localRanks * sizeof(void *));
-  if (!hostPeerPtrs) {
-    res = flagcxSystemError;
-    goto fail;
-  }
-  memset(hostPeerPtrs, 0, localRanks * sizeof(void *));
-
-  for (int lr = 0; lr < localRanks; lr++) {
-    int globalR = localRankToRank[lr];
-    if (globalR == myRank) {
-      hostPeerPtrs[lr] = buff;
-    } else {
-      struct flagcxP2pIpcDesc *pd = &allDescs[globalR];
-      if (pd->size > 0 && deviceAdaptor->ipcMemHandleOpen) {
-        void *peerPtr = nullptr;
-        flagcxIpcMemHandle_t handlePtr = (flagcxIpcMemHandle_t)&pd->handleData;
-        res = deviceAdaptor->ipcMemHandleOpen(handlePtr, &peerPtr);
-        if (res != flagcxSuccess) {
-          WARN("buildIpcPeerPointers: ipcMemHandleOpen failed for rank %d",
-               globalR);
-          goto fail;
-        }
-        hostPeerPtrs[lr] = peerPtr;
-      } else {
-        hostPeerPtrs[lr] = nullptr;
-      }
-    }
-  }
-
-  free(allDescs);
-  allDescs = nullptr;
-
-  // Step 4: Build device-side peer pointer array
-  FLAGCXCHECKGOTO(deviceAdaptor->deviceMalloc((void **)&devPeerPtrs,
-                                              localRanks * sizeof(void *),
-                                              flagcxMemDevice, NULL),
-                  res, fail);
-  FLAGCXCHECKGOTO(deviceAdaptor->deviceMemcpy(
-                      devPeerPtrs, hostPeerPtrs, localRanks * sizeof(void *),
-                      flagcxMemcpyHostToDevice, NULL, NULL),
-                  res, fail);
-
-  // Store in comm->ipcTable
-  comm->ipcTable[slot].hostPeerPtrs = hostPeerPtrs;
-  comm->ipcTable[slot].devPeerPtrs = devPeerPtrs;
-  comm->ipcTable[slot].nPeers = localRanks;
-  comm->ipcTable[slot].basePtr = buff;
-  comm->ipcTable[slot].inUse = true;
-
-  INFO(FLAGCX_INIT,
-       "buildIpcPeerPointers: rank %d slot %d buff=%p devPeerPtrs=%p", myRank,
-       slot, buff, (void *)devPeerPtrs);
-  for (int lr = 0; lr < localRanks; lr++) {
-    INFO(FLAGCX_INIT, "buildIpcPeerPointers:   hostPeerPtrs[%d]=%p", lr,
-         hostPeerPtrs[lr]);
-  }
-  return slot;
-
-fail:
-  free(allDescs);
-  if (hostPeerPtrs) {
-    for (int i = 0; i < localRanks; i++) {
-      if (hostPeerPtrs[i] && hostPeerPtrs[i] != buff) {
-        deviceAdaptor->ipcMemHandleClose(hostPeerPtrs[i]);
-      }
-    }
-    free(hostPeerPtrs);
-  }
-  if (devPeerPtrs) {
-    deviceAdaptor->deviceFree(devPeerPtrs, flagcxMemDevice, NULL);
-  }
-  return -1;
-}
-
-// ==========================================================================
 // Platform wrappers for host memory registration
 // ==========================================================================
