+++ flagcx/adaptor/device_api/nccl_dev_api_backend.cc
@@ -47,18 +47,42 @@ static flagcxResult_t ncclDevApiCommDestroy(flagcxComm_t comm,
 static flagcxResult_t ncclDevApiMemCreate(flagcxComm_t comm, void *buff,
                                           size_t size, flagcxWindow_t win,
                                           flagcxDevMem_t handle) {
-  (void)comm;
   (void)buff;
   (void)size;
-  (void)win;
-  (void)handle;
+
+  // On Vendor path, we only need the ncclWindow_t from the vendor window.
+  // No IPC peer pointers needed — NCCL GIN handles all transport.
+  if (comm != nullptr) {
+    handle->intraRank = comm->localRank;
+  }
+
+  if (win != nullptr && !win->isSymmetricDefault && win->vendorBase) {
+    handle->hasWindow = true;
+    handle->isSymmetric = (win->winFlags & FLAGCX_WIN_COLL_SYMMETRIC) != 0;
+    handle->winHandle = (void *)win;
+  }
+
+  // Allocate and populate kernel Window (wraps ncclWindow_t)
+  auto *kWin = new (std::nothrow) typename DeviceAPI::Window{};
+  if (kWin == nullptr) {
+    return flagcxSystemError;
+  }
+  kWin->populateFromHost(win, handle->rawPtr, handle->intraRank,
+                         handle->mrIndex, handle->mrBase, handle->ipcIndex,
+                         nullptr);
+  handle->window = kWin;
+  handle->hasWindow = kWin->hasAccess();
+
   return flagcxSuccess;
 }
 
 static flagcxResult_t ncclDevApiMemDestroy(flagcxComm_t comm,
                                            flagcxDevMem_t devMem) {
   (void)comm;
-  (void)devMem;
+  if (devMem != nullptr && devMem->window != nullptr) {
+    delete static_cast<typename DeviceAPI::Window *>(devMem->window);
+    devMem->window = nullptr;
+  }
   return flagcxSuccess;
 }
